+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
  - Added WebPortResponse to help IWebPort return byte arrays and headers instad of just strings.
+ - MockIoPort sets configuration and digital out
  
 ### Changed
  - Fixed CongureProxySettings allowing to make more than one request.